strata/requirements.txt
fastapi
uvicorn
celery
redis
httpx
pdfplumber
exa-py
parallel-web
anthropic
supabase
python-dotenv
pydantic
